+++ common/src/main/java/io/github/flemmli97/debugutils/network/PacketRegistrar.java
@@ -11,6 +11,7 @@ public class PacketRegistrar {
 
     public static int registerClientPackets(ClientPacketRegister register, int id) {
         register.registerMessage(id++, S2CDebugToggle.ID, S2CDebugToggle.class, S2CDebugToggle::write, S2CDebugToggle::read, S2CDebugToggle::handle);
+        register.registerMessage(id++, S2CSpawnChunk.ID, S2CSpawnChunk.class, S2CSpawnChunk::write, S2CSpawnChunk::read, S2CSpawnChunk::handle);
         return id;
     }