+++ forge/src/main/java/io/github/flemmli97/debugutils/forge/DebugUtilsClient.java
@@ -6,6 +6,6 @@ import net.minecraftforge.client.event.ClientPlayerNetworkEvent;
 public class DebugUtilsClient {
 
     public static void disconnect(ClientPlayerNetworkEvent.LoggingOut event) {
-        RenderBools.onDisconnet();
+        RenderBools.onDisconnect();
     }
 }