+++ Changelog.md
@@ -1,3 +1,8 @@
+DebugUtils 1.0.2
+================
+- Add api to register more debug renderers 
+- Add debug renderer to show the spawn chunks
+
 DebugUtils 1.0.1
 ================
 - Fix mixin refmap missing